search for event series without events if there are no events with that event series
