+++ common/services/prismic/event-series.js
@@ -1,4 +1,5 @@
 // @flow
+import {getEvents} from './events';
 import type {EventSeries} from '../../model/event-series';
 import type {PrismicDocument} from './types';
 import {
@@ -6,6 +7,7 @@ import {
   parseBackgroundTexture,
   parseBody
 } from './parsers';
+import { getDocument } from './api';
 
 export function parseEventSeries(document: PrismicDocument): EventSeries {
   const genericFields = parseGenericFields(document);
@@ -18,4 +20,26 @@ export function parseEventSeries(document: PrismicDocument): EventSeries {
     backgroundTexture: backgroundTexture ? parseBackgroundTexture(backgroundTexture) : null,
     body: body
   };
+}
+
+type EventSeriesProps = {| id: string |}
+export async function getEventSeries(req: Request, {
+  id
+}: EventSeriesProps) {
+  const events = await getEvents(req, {
+    seriesId: id
+  });
+
+  if (events && events.results.length > 0) {
+    const series = events.results[0].series.find(series => series.id === id);
+    return {
+      series,
+      events: events.results
+    };
+  } else {
+    // TODO: (perf) we shoulnd't really be doing two calls here, but it's for
+    // when a series has no events attached.
+    const document = await getDocument(req, id, {});
+    return document && { series: parseEventSeries(document), events: [] };
+  }
 }
