+++ server/middleware/error.js
@@ -3,7 +3,7 @@ import {createPageConfig} from '../model/page-config';
 
 export function error(beaconError) {
   return async (ctx, next) => {
-    const isPreview =
+    const isPreview = true ||
       Boolean(ctx.request.url.match('/preview')) ||
       Boolean(ctx.request.host.match('preview.wellcomecollection.org'));
