+++ whats_on/app/controllers.js
@@ -8,7 +8,8 @@ import {
   getExhibitionExhibits,
   getExhibitExhibition
 } from '@weco/common/services/prismic/exhibitions';
-import { getEvent, getEventSeries } from '@weco/common/services/prismic/events';
+import {getEvent} from '@weco/common/services/prismic/events';
+import {getEventSeries} from '@weco/common/services/prismic/event-series';
 import {getEventbriteEventEmbed} from '@weco/common/services/eventbrite/event-embed';
 import {isPreview as isPrismicPreview} from '@weco/common/services/prismic/api';
 import {model, prismic} from 'common';
